requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
pydantic
python-dotenv